refactor: bump nx version to ^17.2.8
