+++ .prettierignore
@@ -2,3 +2,5 @@
 
 /dist
 /coverage
+
+/.nx/cache
