+++ package.json
@@ -18,46 +18,48 @@
   "devDependencies": {
     "@commitlint/cli": "^16.2.3",
     "@commitlint/config-angular": "^16.2.3",
-    "@forge/cli": "^6.9.1",
-    "@nx/devkit": "16.1.4",
-    "@nx/eslint-plugin": "16.1.4",
-    "@nx/jest": "16.1.4",
-    "@nx/js": "16.1.4",
-    "@nx/linter": "16.1.4",
-    "@nx/node": "16.1.4",
-    "@nx/plugin": "16.1.4",
-    "@nx/webpack": "16.1.4",
-    "@nx/workspace": "16.1.4",
-    "@swc-node/register": "^1.4.2",
-    "@swc/core": "^1.2.173",
-    "@types/fs-extra": "^9.0.13",
-    "@types/jest": "29.4.4",
-    "@types/js-yaml": "^4.0.5",
-    "@types/node": "18.7.1",
-    "@typescript-eslint/eslint-plugin": "5.59.1",
-    "@typescript-eslint/parser": "5.59.1",
-    "dotenv": "^16.0.3",
-    "eslint": "8.15.0",
-    "eslint-config-prettier": "8.1.0",
-    "graphql": "^16.6.0",
-    "graphql-request": "^6.0.0",
-    "husky": "^7.0.4",
+    "@nx/devkit": "17.2.8",
+    "@nx/eslint-plugin": "17.2.8",
+    "@nx/jest": "17.2.8",
+    "@nx/js": "17.2.8",
+    "@nx/node": "17.2.8",
+    "@nx/plugin": "17.2.8",
+    "@nx/webpack": "17.2.8",
+    "@nx/workspace": "17.2.8",
+    "@swc-node/register": "1.6.7",
+    "@swc/core": "1.3.103",
+    "@types/fs-extra": "9.0.13",
+    "@types/jest": "29.5.11",
+    "@types/js-yaml": "4.0.9",
+    "@types/node": "18.16.9",
+    "@typescript-eslint/eslint-plugin": "6.19.0",
+    "@typescript-eslint/parser": "6.19.0",
+    "dotenv": "16.3.1",
+    "eslint": "8.48.0",
+    "eslint-config-prettier": "9.1.0",
+    "graphql": "16.8.1",
+    "graphql-request": "6.1.0",
+    "husky": "^8.0.3",
     "jest": "29.4.3",
     "jest-environment-jsdom": "29.4.3",
     "jest-environment-node": "^29.4.1",
-    "nx": "16.1.4",
+    "nx": "17.2.8",
     "prettier": "2.8.0",
     "semantic-release": "^19.0.2",
     "ts-jest": "29.1.0",
     "ts-node": "10.9.1",
     "tslib": "^2.3.0",
-    "typescript": "5.0.4",
-    "zod": "^3.21.4"
+    "typescript": "5.2.2",
+    "zod": "^3.21.4",
+    "@nx/eslint": "17.2.8"
+  },
+  "peerDependencies": {
+    "@forge/cli": "^7.0.0"
   },
   "dependencies": {
     "@fastify/autoload": "~5.7.1",
     "@fastify/sensible": "~5.2.0",
-    "@forge/manifest": "^4.4.1",
+    "@forge/manifest": "5.5.0",
     "axios": "^1.0.0",
     "fastify": "~4.13.0",
     "fastify-plugin": "~4.5.0",
@@ -70,4 +72,3 @@
     "url": "https://github.com/toolsplus/nx-forge.git"
   }
 }
-
