+++ packages/nx-forge/package.json
@@ -6,9 +6,9 @@
   "generators": "./generators.json",
   "executors": "./executors.json",
   "peerDependencies": {
-    "@nx/webpack": "^16.0.0",
-    "@nx/workspace": "^16.0.0",
-    "@nx/node": "^16.0.0"
+    "@nx/webpack": "^17.2.8",
+    "@nx/workspace": "^17.2.8",
+    "@nx/node": "^17.2.8"
   },
   "nx-migrations": {
     "migrations": "./migrations.json"
