+++ packages/nx-forge/src/generators/application/generator.ts
@@ -6,9 +6,8 @@ import {
   runTasksInSerial,
   joinPathFragments,
 } from '@nx/devkit';
-import { Linter, lintProjectGenerator } from '@nx/linter';
+import { Linter, lintProjectGenerator } from '@nx/eslint';
 import { jestProjectGenerator } from '@nx/jest';
-import { mapLintPattern } from '@nx/linter/src/generators/lint-project/lint-project';
 import initGenerator from '../init/generator';
 import { ApplicationGeneratorOptions } from './schema';
 import { addProject, createFiles, normalizeOptions } from './lib';
@@ -36,9 +35,9 @@ export default async function (
       tsConfigPaths: [
         joinPathFragments(options.appProjectRoot, 'tsconfig.app.json'),
       ],
-      eslintFilePatterns: [
-        mapLintPattern(options.appProjectRoot, options.js ? 'js' : 'ts'),
-      ],
+      // eslintFilePatterns: [
+      //   mapLintPattern(options.appProjectRoot, options.js ? 'js' : 'ts'),
+      // ],
       unitTestRunner: options.unitTestRunner,
       skipFormat: true,
       setParserOptionsProject: options.setParserOptionsProject,
