+++ packages/nx-forge/src/generators/application/lib/normalize-options.ts
@@ -1,4 +1,4 @@
-import { Linter } from '@nx/linter';
+import { Linter } from '@nx/eslint';
 import type { Tree } from '@nx/devkit';
 import {
   extractLayoutDirectory,
